splitting code, separating stateless functional components, creating filters selects
